Added comments to Graph class Created a "validateStation" method in RailwayController class to existence check the input station Added some user input validation for finding an accessible route Added a getEdgeWeight method to Graph.
Added a calculateBoardingTime method to RailwayController that has the skeleton for finding the departure time. Takes in a string, validates it looks like a 24hr format time, converts it to minutes, and then converts it back to 24hr time.

Tiedied up the RailwayController class.

Added some more comments and javadoc method descriptions.